printer_agent/requirements.txt
websockets>=12.0
requests>=2.28
pystray>=0.19
Pillow>=10.0
pywin32>=306; sys_platform == "win32"
pyinstaller>=6.0; sys_platform == "win32"
